refactor(iam): reorganize IAM command structure and enhance retry logic for various commands
