+++ src/commands/iam/permissions/get-user-permissions.ts
@@ -39,7 +39,7 @@ export class UserPermissionsCommand extends Command<
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
