+++ src/commands/iam/policy-associations/get-key-policies.ts
@@ -85,7 +85,7 @@ export class KeyPoliciesCommand extends Command<KeyPoliciesInput, Policy[]> {
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
