+++ src/commands/iam/policy-associations/get-organization-policies.ts
@@ -20,7 +20,7 @@ export class OrganizationPoliciesCommand extends Command<
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
