+++ src/commands/iam/policy-associations/get-policy-associations.ts
@@ -143,7 +143,7 @@ export class PolicyAssociationsCommand extends Command<
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
