+++ src/commands/iam/policy-associations/get-role-policies.ts
@@ -69,7 +69,7 @@ export class RolePoliciesCommand extends Command<RolePoliciesInput, Policy[]> {
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
