+++ src/commands/iam/policy-associations/get-user-policies.ts
@@ -19,7 +19,7 @@ export class UserPoliciesCommand extends Command<UserPoliciesInput, Policy[]> {
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
