+++ src/commands/iam/role-associations/get-organization-roles.ts
@@ -33,7 +33,7 @@ export class OrganizationRolesCommand extends Command<
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
