+++ src/commands/iam/role-associations/get-role-associations.ts
@@ -71,7 +71,7 @@ export class RoleAssociationsCommand extends Command<
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
