+++ src/commands/iam/roles/get-roles.ts
@@ -17,7 +17,7 @@ export class RoleListCommand extends Command<RoleListInput, Role[]> {
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
