+++ src/commands/iam/roles/id/get-role.ts
@@ -16,7 +16,7 @@ export class RoleGetCommand extends Command<RoleGetInput, Role> {
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
