+++ src/commands/iam/validate/validate-key.ts
@@ -28,7 +28,7 @@ export class ValidateKeyCommand extends Command<
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
