+++ src/commands/iam/validate/validate-user.ts
@@ -70,7 +70,7 @@ export class ValidateUserCommand extends Command<
   /**
    * Whether the command should retry on failure
    */
-  protected override retryOnFailure = false
+  protected override retryOnFailure = true
 
   /**
    * Get the method
